simple_cms_api/requirements.txt
Django==3.0.7
djangorestframework==3.11.0
psycopg2==2.8.5
boto3==1.14.17
django-storages==1.9.1
django-oauth-toolkit==1.3.2
django-cors-headers==3.4.0